refactor: "패키지 변경 (controller 내의 console 폴더)"
