--- src/main/java/org/prgrms/prgrmsspring/controller/ApplicationController.java
@@ -1,5 +1,0 @@
-package org.prgrms.prgrmsspring.controller;
-
-public interface ApplicationController {
-
-}
