+++ src/main/java/org/prgrms/prgrmsspring/controller/console/ApplicationController.java
@@ -1,0 +1,5 @@
+package org.prgrms.prgrmsspring.controller.console;
+
+public interface ApplicationController {
+
+}
