+++ src/main/java/org/prgrms/prgrmsspring/domain/ControllerManager.java
@@ -4,10 +4,10 @@ import org.prgrms.prgrmsspring.console.ConsoleIOManager;
 import org.prgrms.prgrmsspring.console.CustomerConsole;
 import org.prgrms.prgrmsspring.console.VoucherConsole;
 import org.prgrms.prgrmsspring.console.WalletConsole;
-import org.prgrms.prgrmsspring.controller.ApplicationController;
-import org.prgrms.prgrmsspring.controller.CustomerController;
-import org.prgrms.prgrmsspring.controller.VoucherController;
-import org.prgrms.prgrmsspring.controller.WalletController;
+import org.prgrms.prgrmsspring.controller.console.ApplicationController;
+import org.prgrms.prgrmsspring.controller.console.CustomerController;
+import org.prgrms.prgrmsspring.controller.console.VoucherController;
+import org.prgrms.prgrmsspring.controller.console.WalletController;
 import org.prgrms.prgrmsspring.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
